+++ library/src/commonMain/kotlin/io/github/ronjunevaldoz/paymongo/IPayMongo.kt
@@ -1,6 +1,21 @@
 package io.github.ronjunevaldoz.paymongo
 
-import io.github.ronjunevaldoz.paymongo.models.resource.*
+import io.github.ronjunevaldoz.paymongo.models.resource.AttachPaymentIntentInput
+import io.github.ronjunevaldoz.paymongo.models.resource.CheckoutSessionResponse
+import io.github.ronjunevaldoz.paymongo.models.resource.CreateCheckoutSessionInput
+import io.github.ronjunevaldoz.paymongo.models.resource.CreateLinkInput
+import io.github.ronjunevaldoz.paymongo.models.resource.CreatePaymentInput
+import io.github.ronjunevaldoz.paymongo.models.resource.CreatePaymentIntentInput
+import io.github.ronjunevaldoz.paymongo.models.resource.CreatePaymentMethodInput
+import io.github.ronjunevaldoz.paymongo.models.resource.CreateSourceInput
+import io.github.ronjunevaldoz.paymongo.models.resource.CreateWebhookInput
+import io.github.ronjunevaldoz.paymongo.models.resource.PaymentIntentResponse
+import io.github.ronjunevaldoz.paymongo.models.resource.PaymentLink
+import io.github.ronjunevaldoz.paymongo.models.resource.PaymentMethodResponse
+import io.github.ronjunevaldoz.paymongo.models.resource.PaymentResponse
+import io.github.ronjunevaldoz.paymongo.models.resource.SourceResponse
+import io.github.ronjunevaldoz.paymongo.models.resource.WebhookResponse
+import io.github.ronjunevaldoz.paymongo.models.resource.WebhooksResponse
 
 
 /**
@@ -9,7 +24,6 @@ import io.github.ronjunevaldoz.paymongo.models.resource.*
 interface IPayMongo {
     // source
     suspend fun createSource(input: CreateSourceInput): SourceResponse
-
     /**
      *  @param [id] Source id
      */
@@ -29,12 +43,18 @@ interface IPayMongo {
     /**
      *  @param [paymentIntentId] PaymentIntent id
      */
-    suspend fun getPaymentIntent(paymentIntentId: String, clientKey: String? = null): PaymentIntentResponse
+    suspend fun getPaymentIntent(
+        paymentIntentId: String,
+        clientKey: String? = null
+    ): PaymentIntentResponse
 
     /**
      *  @param [paymentIntentId] PaymentIntent id
      */
-    suspend fun attachPaymentIntent(paymentIntentId: String, input: AttachPaymentIntentInput): PaymentIntentResponse
+    suspend fun attachPaymentIntent(
+        paymentIntentId: String,
+        input: AttachPaymentIntentInput
+    ): PaymentIntentResponse
 
     // webhooks
     suspend fun createWebhook(input: CreateWebhookInput): WebhookResponse
@@ -56,4 +76,15 @@ interface IPayMongo {
     suspend fun createCheckoutSession(input: CreateCheckoutSessionInput): CheckoutSessionResponse
     suspend fun getCheckoutSession(checkoutSessionId: String): CheckoutSessionResponse
     suspend fun expireCheckoutSession(checkoutSessionId: String): CheckoutSessionResponse
+
+    /**
+     * New!
+     * Create a Link
+     * https://developers.paymongo.com/reference/create-a-link
+     */
+    suspend fun createLink(input: CreateLinkInput): PaymentLink
+    suspend fun getLink(id: String): PaymentLink
+    suspend fun getLinkByReference(referenceNumber: String): PaymentLink
+    suspend fun archiveLink(id: String): PaymentLink
+    suspend fun unarchiveLink(id: String): PaymentLink
 }