+++ samples/demo/src/wasmJsMain/kotlin/main.kt
@@ -1,8 +1,6 @@
-import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.window.ComposeViewport
 import kotlinx.browser.document
-import kotlinx.browser.window
 
 @OptIn(ExperimentalComposeUiApi::class)
 fun main() {