+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [v1.0.2-dev01]
+
+### Added
+- Create a link
+- Get link by id
+- Get link by reference number
+- Archive link
+- Unarchive link
+- Added new webhook `Event` `link.payment.paid` and `checkout_session.payment.paid`
+
 ## [v1.0.1]
 
 ### Changed